Move import sink to specific ExecImportSink
Summary: User controlled input flowing into import statements doesn't directly lead to a RCE vulnerability but can be chained to another issue to create a full exploit. Moving this to a separate sink to distinguish this case from code, code injection issues. The risk is similar to a user controlled environment variable merging this 2 cases in the same category

Reviewed By: gbleaney

Differential Revision: D27649317

fbshipit-source-id: c48d47040e66829ca89c83451849325e6bee15fd