Generalize Seq
Summary:
Seq(op1, op2) before was limited to op1 being a transform. This was not really useful in reductions or partitions, since such transforms can be included in the op2 function directly.

Seq(op1, op2) now allows op1 and op2 to be both transforms, or both reductions. For reductions, this allows reducing over multiple parts in a single traversal.

For partitions, this would mean partitioning to the cross-product. Unfortunately, our operation types cannot express this as the return type is then distinct from the transform/reduce cases. (We could add a CoPartition operation if this is ever needed).

Reviewed By: arthaud

Differential Revision: D27633329

fbshipit-source-id: c4e5d313901513eee59a1592ce5307ace3e59df3