Update installing dependencies instructions for Ubuntu, Mint, and Debian
Summary:
# watchman
The current docs mention you can install `watchman` with `sudo apt-get install watchman`. But from my testing so far, it doesn't seem like Ubuntu 18.04.5 LTS can locate `watchman` with the default list of repositories that comes with Ubuntu.

Also, the [official watchman docs](https://facebook.github.io/watchman/docs/install.html#installing-on-macos-or-linux-via-homebrew) don't mention installing watchman through `apt-get`. Although that don't necessarily indicate that it is not possible, but it seems like the recommended way is with homebrew

# python3-venv
We are missing `python3-venv` in ` sudo apt-get install python3 python3-pip`, so users will not be able to follow the instructions to create a virtual environment in the "Setting up a Project" section

Reviewed By: dkgi

Differential Revision: D27685692

fbshipit-source-id: 6efba33e02b92553607a7b61c2b1fc8e0819c3a4