Special handle for DataFrame getitem and setitem
Summary:
Due the complexity of pandas dataFrame api, we decide to specially handle some pandas api when lineage-analysis is true.
In this diff, we handle two api: getitem and setitem for dataFrames.
For getitem, the input can be a string or a list of string.
If the input is a string, it returns a series.
If the input is a list of string, it returns a dataFrame that contains corresponding series.
For setitem, the input index can also be a string or a list of string.
If the input is a string, the input value is a Series. It sets that index is corresponding to that Series.
If the input is a list of string, the input value is a DataFrame. It sets each index in the inputs, equals to the corresponding series of the input DataFrame.

Reviewed By: arthaud

Differential Revision: D27602609

fbshipit-source-id: 705d7de47f0c6741cb716bfa50bedc698188fc76