Restructure `analyze` slightly for readability
Summary:
For interprocedural infer, I want to rework the externalization
of analyses so that more of the logic is pushed down to the
ANALYSIS_PROVIDED level.

Whenever I read this code though, I was getting slightly confused
by the fact that `save_results` was actually defined out-of-order
in order to reuse it in the try/with. It's not hard to
understand the code as it currently is, but it's a little hard to *skim*.

I find this easier: define run_analysis before save_result,
then use both of them in a try/with.

Reviewed By: arthaud

Differential Revision: D27773192

fbshipit-source-id: dc56e5c1d9ff7fd120517cfed12e6a3617a9e6bd