Refactor to avoid nested tuples
Summary:
At least to me, it is pretty challenging to unpack the way that
we were dealing with the data from `access_as_attribute`.

This diff refactors so that we never have doubly-nested tuples,
by deferring actually pulling an annotation out of the attribute
until we need it, instead of doing it early inside `access_as_attribute`.

The one thing that's worse in this change is we have less type safety
around knowing that `annotations` is non-empty, but I think the
improved clarity is worth it; we call `List.hd_exn` only a few lines
below where we pattern-matched a nonempty list.

Here's what was happening on master:
- `access_as_attribute` was returning a doubly-nested tuple
- we were unpacking the tuples, but the `definition` parts of the
  partially-unpacked tuples were still tuples
- at one stage, we were splitting appart the nested `definitions`
  topules and then *re-packing* part of the data with a different part
  of the outer tuple(!!)
  - also the variable name `definition` was getting used as a catch-all
    for any tuple we didn't know how to actually give a usedful name,
    which further added to the confusion

My guess is the motivation for doing all this is that there's no
`List.unzip4` function, but there's actually no reason to convert
attributes to annotations ahead of time.

Reviewed By: grievejia

Differential Revision: D27823997

fbshipit-source-id: e00e806266253b3ceef068f6b77b709f6863e8d4