Distinguish variable names more in attribute resolution
Summary:
In the context of resolving values in `typeCheck.ml`, the name
`resolved` is almost always used for a `Resolved.t`.

But in the attribute resolution code, the same variable name was
being used for, alternatively,
- a `Resolved.t`, which is reasonable
- a class_data record
- a resolved annotation coming from global attribute resolution
- a resolved annotation coming from local overrides

This made it much harder to skim the code in my opinion than if
we use different, descriptive names for each of these things.

So this commit
- calls class_data by the name class_data
- calls annotations by the name annotation, and where both have to
  be in scope at the same time distinguishes between global_annotation and
  local_annotation
- calls only the very final annotation resolved_annotation... anything up
  to that point isn't really resolved yet

Also in the process I renamed `apply_global_override`, to
`apply_local_override`. You could argue for either name making sense, since
it's applying a local override to a global annotation, but after taking into
account the fact that we need to have reasonable names for both the input and
output variables I think it makes sense to emphasize that the override itself
is local, the *argument* is the default, global annotation.

Reviewed By: shannonzhu

Differential Revision: D27822765

fbshipit-source-id: 56e2e557dc452407a8022fee68e1d85be39f56d4