Add support for negating model constraint clauses
Summary:
As titled. This adds support for a `Not` clause used to negate existing model constraint clauses, allowing us to specify things like not matching a regex, or not subclassing a particular class, etc. e.g.

```
ModelQuery(
  find = "methods",
  where = [
    parent.matches("model_query.ClassTest1"),
    parent.matches("media\.urlgen\.MediaURLGenerationMixin"),
    name.matches(".*_url.*"),
    Not(name.matches(".*as_slipstream.*"))
  ],
  model = [
    Returns(TaintSource[SourceB]),
  ]
)
```

It's safe to negate the `matches_attribute_constraint` function, since the diff above this one in the stack ensures that we don't reach an error case with an invalid clause in here. For `matches_callable_constraint`, there might be one error case where `callable_type` is `None`, but I ran on IG with some logging added and didn't see this occur. I'm not sure when this occurs, so I'll just keep the previous behavior for now and add some logging in case this does ever occur.

Reviewed By: arthaud

Differential Revision: D27824810

fbshipit-source-id: 5729b7489ae15e18f0afbc56ddd0e947d0826b7b