Rearrange code in interproceduralAnalysis.ml
Summary:
I'm working on a refactor of interprocedural results handling, and it's
easier to visualize the code I'm refactoring if all of the result-extraction
logic lives next to one another.

It's also easier to skim the code if it's laid out in the actual order
steps get taken:
- initialization: `initialize` and `record_initial_models`
- fixpoint computation: everything from there till `compute_fixpoint`
- result extraction: `externalize`, `extract_errors`, and `save_results`

This commit is a pure rearrange of the `.ml` and `.mli` file to follow this
principle.

Reviewed By: shannonzhu

Differential Revision: D28019175

fbshipit-source-id: 0532b6e52d1e0a1203be425078d71163c63ed825