Add `is_transitive` flag to parent.extends
Summary:
This gives us the ability to have multiple levels of inheritance for `parent.extends`.

For example, given:
```
class C:
   ...

class D(C):
   ...

class E(D):
   ...
```

A query like:
```
ModelQuery(
  find = "methods",
  where = [parent.extends("C", is_transitive=True)],
  model = [
    Returns(TaintSource[Test])
  ]
)
```

will taint all the methods of `C`, `D` and `E`.

Note that the base class is tainted; an additional `Not(parent.matches("C"))` in the where clause would be necessary to exclude tainting the base class, to get the behavior of our existing Python model generators.

Reviewed By: pradeep90

Differential Revision: D27946875

fbshipit-source-id: c5390c772beb99372ff5cd12ed97299cd51964bc