Interprocedural: refactor results handling [1] move error handling
Summary:
I'm preparing to modify the interprocedural framework so that all of the
results surfacing logic gets dealt with in an Analysis-controlled way.

As part of doing this, I really want to also unify the error handling, which
at this point is not easy to see, so that it lives in one place and so that
new interprocedural analyses don't have to implement interfaces that are only
there for taint-specific tests and debugging.

This commit is a first step: move the error printing from analyzeCommand
into the results-externalization logic at then endo of `StaticAnalysis.analyze`.

My goal is to keep pushing it further down until it lives in taint-specific
code only, so that we can remove the `get_errors` interface entirely and just
have a single (global) externalize function in ANALYSIS_PROVIDED

Reviewed By: arthaud

Differential Revision: D28005527

fbshipit-source-id: fb1492d6a29573f0643ecdb597768e7d21884bf4