Interprocedural: refactor results handling [2] move report_results
Summary:
This commit moves the newly-extracted `report_results`
function out of `StaticAnalysis` and into `InterproceduralAnalysis`, without
changing any of the logic.

This move is needed because I'm working on making saving and printing all
happen at a per-analysis level in a way that the ANALYSIS_PROVIDED can control,
and that means that I need to keep pushing this code down the callstack until
it's called on a per-abstract-analysis basis.

The next diff will rewrite parts of `report_results` and
`save_results` to accomplish this, which will ultimately produce our concrete
implementation of the taint externalization.

Reviewed By: arthaud

Differential Revision: D28008898

fbshipit-source-id: 0783deccccb482bfdbcaedf493b79287b6ceeb59