Rename Taint.Configuration -> Taint.TaintConfiguration
Summary:
I'm working toward allowing interprocedural analyses to completely control
their own reporting, which is needed for type inference because the pysa-style
reporting structure isn't well suited to type inference.

The taint reporting function I'll need has to consume a pyre
Configuration.StaticAnalysis.t record, and it turns out that due to module
name resolution rules a hard blocker on this is to not have a
taint-specific Configuration module, which prevents me from accessing the
top-level Configuration module.

This seems like a good change anyway - it's not as clear to have the
taint-specific configuration sometimes called Configuration and other times
TaintConfiguration as it is to just always call it TaintConfiguration.

Reviewed By: arthaud

Differential Revision: D28044274

fbshipit-source-id: 2aadf3071e315932954b3b567d2e7a6ea8acd0c2