Fix qualified name of helper functions.
Summary:
Problem: We were storing the decorator module name using the original name of the helper. That name was fully-qualified whereas the decorator-storing function expected just the last part of the name. This meant the filename for helper functions was wrong.

Use the last part of the helper function name.

Reviewed By: arthaud

Differential Revision: D28047976

fbshipit-source-id: 281db4f8884f87caeb373477000b0fb4bc22e3d6