Return decorator module location along with definitions.
Summary:
Problem: When we inline a decorator, we end up showing its traces on line numbers corresponding to its original file but on the decorated function's file.

We want to show those lines on the original file. Pysa deals with `Location.WithModule` in its models and issues and converts them to `Location.WithPath` only when emitting json. So, we have to  emit `decorator_module` instead of `decorated_function_module` in the issues and models for the inlined decorator functions.

This diff gets the decorator module when collecting all known decorators.

Note: I suspect it should be enough to do `Reference.prefix` instead of getting the module location from the environment, but doing it anyway just to be safe.

Reviewed By: arthaud

Differential Revision: D28043155

fbshipit-source-id: 98d23b7319d588d19f2642d942c278c080327bab