Point to decorator call instead of dynamic return statement.
Summary:
Problem: Zoncolan UI shows traces against source code from a given commit. However, we inline decorators dynamically, which means that our inlined function doesn't correspond directly to the code provided.

There are four ways in which the inlined function differs from the original function:
1. `foo.original_function`
2. `foo.wrapper_function`
3. `foo.some_helper`
4. `return foo.wrapper(x, y)`

Our design keeps `foo.original_function` and `foo.wrapper_function` as first-class functions within the inlined function. The advantage of doing this is that their locations correspond exactly to the original function and the decorator wrapper.

There are two problems:
+ The main problem is that the line number is that of the decorator but the filename is that of the decoratee. We can't change this while inlining because the filename is calculated automatically based on the module in which the function is found. Will address this in upcoming diffs.
+ (4) was pointing at nonsensical locations because the location was relative to the decorator file.

This diff makes (4) point to the `with_logging` decorator call. That way, users can understand that the flow is through the decorator `with_logging` and the function call is to the wrapper within `with_logging`.

Tradeoff: There will be a discrepancy here because the displayed line will be `with_logging` whereas the source line will actually be `return __wrapper`. I don't know of a good solution here since there will always be a dynamically-generated line that does not correspond to any original line. I think this will have to be a compromise here unless there are better suggestions for showing this.

Reviewed By: arthaud

Differential Revision: D28043158

fbshipit-source-id: b0bf9e1b8875be34384d21483b16c5ce667dbd86