Store mapping from inlined function to decorator module.
Summary:
Problem: We want to show the inlined decorator function in its original file on the Zoncolan UI. However, it is currently treated as being part of the decorated function's module.

So, we will store the decorator's module when inlining it. Doing this in shared memory because there will be different workers inlining decorators.

Yet to use this.

Reviewed By: arthaud

Differential Revision: D28047977

fbshipit-source-id: 47659dda3d9c27a8f9cc5f36ff507c5c476f4a1d