Fix client error when --save-results-to is used for pysa
Summary:
There are no actual integration tests that cover the production pysa
command, which uses --save-output-to to dump output as json.

As a result, I failed to notice that in D28044272 (https://github.com/facebook/pyre-check/commit/687b27581bff12329b9ddbddc0ad8d00b4e7639a) I caused the client to
break, it logs an error
```
2021-04-28 14:41:49,606 [PID 837654] ERROR Invalid JSON output: ``.
```
because I'm not pushing "[]" to stdout anymore, and the client still tries
to parse stdout as a json list.

This commit fixes the problem - I log a json empty list in the case when
we're dumping output to a file.

Reviewed By: arthaud

Differential Revision: D28076158

fbshipit-source-id: 7937c6495b571bdf84b845e6fb1a9461982b42e8