Do not count modules ignoring specific error codes as explicitly unsafe
Summary:
In cases where a project has turned strict default on, we were still counting modules that had a file-level suppression of an error code as not strict.

This was raised as a point of frustration for project owners that were pushing for 100% strict in their project, but were blocked on noisy errors like `# pyre-fixme[56]` firing one very application of a decorator inside a module, which was made the code pretty unreadable when suppressed line by line.

The module is still strict in the sense that every global, attribute, and function in that module has to be annotated, etc.

There probably should not be a tradeoff between being able to push for strict adoption and being able to suppress errors of a specific error code. Of course, ignore-all-errors still counts as an explicit unsafe file.

I also have a task T84083316 to support disallowing ignore-all-errors from any file in your project, when that becomes the coverage issue at hand to tackle.

----

Reviewed By: dkgi

Differential Revision: D28082969

fbshipit-source-id: e93212a2191b90518b2bc8cad99f4cfadc5a7bab