Fix subtraction in product domain
Summary:
Product domains with strict fields acted badly in subtraction. E.g., a product domain of 2 sets, where the left side is strict in bottom cannot collapse to bottom during subtraction.

a \ b = c means c <= a and a <= b |_| c

({a}, {b,c}) \ ({a}, {b})  cannot be equal to bottom ({}, {}) even if it is strict in the left, as the a <= b |_| c inequality won't hold on the result.

Thus, when doing subtraction on products with strict slots, we have to be more careful. We cannot collapse. In case a strict slot is bottom, we have to keep the original value in that slot, unless all slots are bottom.

Separate small fix for strict maps. `of_list` wasn't strict-conform

Reviewed By: arthaud

Differential Revision: D28039538

fbshipit-source-id: 46e831b399a01c7693f087538f8f0c5b97368c46