Interprocedural: refactor results handling [7] simplify ANALYSIS_PROVIDED
Summary:
This commit follows up adding `report` to the `ANALYZER` interface with
removing a bunch of reporting-related functions from the `ANALYSIS_PROVIDED`
interface: `get_errors`, `metadata`, `statistics`, and `externalize`.

Obviously any analysis is still free to use these functions, but given that
we now control the entire report there's no reason for the interprocedural
framework to include them in the interface.

Reviewed By: arthaud

Differential Revision: D28046237

fbshipit-source-id: c4121093f55b70db4464e88615ecf205cfba2b46