Interprocedural: refactor results [6] `report` -> ANALYZER
Summary:
This commit is the main cutover that all previous commits have been prepping for:
- Add a new `report` function to the ANALYZER interface, which reports
  results globally
  - This in turn required updating a bunch of tests, as well as the stubbed-out
    typeInferenceAnalysis and deadStoreAnalysis modules
- Move the `report_analysis` method from `interproceduralAnalysis.ml` to a new
  module `TaintReporting`, and hook that up to the `TaintAnalysis.report`, along
  with all the helper functions

In the process, I was able to remove quite a lot of the complexity (a bit of
code code, and a lot of confusing type gymnastics) from the externalize and
error-fetching functions, because they no longer have to be generic over
Kind.abstract values.

We can now remove the `externalize` and `get_errors` functions
from the `ANALYSIS_PROVIDED` interface and clean up a bunch of stubbed code.
I decided not to do that in this commit because it's probably easier to review adding
and removing interface logic in separate commits, but I'll get rid of all the cruft in
my next commit.

Reviewed By: arthaud

Differential Revision: D28044273

fbshipit-source-id: 9d94f82604f46a0c9fb2730e675ae84d50213fda