Call Configuration.StaticAnalysis static_analysis_configuration
Summary:
We frequently need to have in scope both a Configuration.StaticAnalysis
and a Configuration.Analysis at the same time.

To make naming more consistent and skimming easier, rename all interprocedural
code to use `static_analysis_configuration` as the name for the former,
and `configuration` as the name for the latter (which is also the field in
the former for an analysis config).

Reviewed By: grievejia

Differential Revision: D28105647

fbshipit-source-id: 78fd0c4350772ff31dcea30e33a7e32016117243