Split the handling of callable and attribute models
Summary:
Right now, the model parser treats attribute/object models the same as callables, by defining a fake signature for the attribute.
This leads to inconsistencies where we accept `def Foo.bar() -> TaintSource[...]` models for attributes, and generate a function model instead of an object model, which ultimately leads to non-determinism in the analysis (see T79327742 for more context).

To fix this, we want to prohibit using `def Foo.bar() -> TaintSource[..]` for attributes. This is very hard with the current code. This patch refactors the model parser to split attribute and callable model parsing.

Reviewed By: dkgi

Differential Revision: D28136210

fbshipit-source-id: cd529cb645532bc7cb858fb7db0d206de555f09c