Support more understandable naming for source dir import root
Summary:
Naming RFC.

Given there are use cases on both ends of this behavior (ie., some use cases want to assume that the source directory given is also the import root for all imports in those files, some use cases want to assume that the pyre configuration directory (or some other path) is the import root - see background context below for examples), there doesn't seem to be a clear-cut case for changing the default.

I'm erring on the side of leaving the default behavior as-is. In other words, a source directory given as a plain string means that this is both the root of the sources we care about, *and* the import root that all imports will be relative to.

I think this makes sense both for backwards compatibility, and relatively intuitive default behavior (as opposed to the source directory specifies the sources, the project root specifies the import root). Suggestions welcome on this decision, though.

If we are keeping the default, I think appropriate steps to make this usable for open source onboarding include:
(1) find an appropriate name that is less confusing than "root"/"subdirectory" which our own team even gets confused about
(2) pyre init should be changed to output a default configuration that conforms with the import root standard: i.e., set the import root as "." by default if the source dir is not ".", perhaps including a prompt to users.

RFC for the naming change:
1. Are the names clear/intuitive?
2. Should we change the search path to use the same names? Deprecate the old root/subdirectory ones?
3. As-is, the name doesn't imply a relationship between the import root and the source path in the same way as "subdirectory" implies that this string is relative to the root just specified. I think this is an improvement in that it wasn't clear what subdirectory or root meant outside of their relative definition to one another, but should we error if the given source root is not a subdirectory of the import root?

----
**Background Context:**

Running pyre on an open source project, in this case using pytorch as an example, with (rooted in pytorch/.pyre_configuration):

`source_directories: ["torch"]`

will result in a lot of import errors if you attempt to import using:

`from torch.a.b import X`

Instead, Pyre expects the above import to be given in the form

`from a.b import X`

The reason we do this historically is because distillery is a collection of a lot of different python packages. Distillery is an aggregate directory that is not itself a package, but includes toplevel directories that are themselves packages.
So in distillery, from `distillery.a.b import X` does not make sense - all imports are of the form `from a.b import X`

Update: Internally, whatsapp repositories also use this format - where they expect imports to be relative to the given source directory, not to the location of the .pyre_configuration. (https://www.internalfb.com/intern/diffusion/WAANDROID/browsefile/master/.pyre_configuration?lines=3)
Even within fbcode, there is some usage of source_directories that expects imports to be relative to the source dir: (https://www.internalfb.com/intern/diffusion/FBS/browse/master/fbcode/social_wifi/.pyre_configuration.local?commit=0d3aa685f304&lines=7)
There are also some cases where setting the import root has been hacked around: (https://www.internalfb.com/intern/diffusion/FBS/browse/master/xplat/whatsapp/wavoip/.pyre_configuration?commit=0d3aa685f304&lines=3)

For most open source cases, the given repository should type check as the root of imports.

The only workaround for this case right now is specifying:

`source_directories: [{"root": ".", "subdirectory": "torch"}]`

Note: sapp also has a `sapp/sapp` folder, but the current setup doesn't suffer from this issue because we specify the source directory as `"."`. It would have the same problem as pytorch if we only cared about type checking `sapp/sapp`, and specified the source directory as `"sapp"`.

Reviewed By: dkgi

Differential Revision: D28101051

fbshipit-source-id: 0be274bfb637ff5e39cb5d90ee2988bb58b81bae