Support multiple directories in pyre init prompt
Summary:
Make this prompt consistent with the one we use to gather source directories for analysis when creating a local configuration.

Currently, the prompt phrasing of "which directory should pyre be initialized in" sounds like we will place the .pyre_configuration file in a different place depending on the answer, when in reality we want to ask which directories should be type checked. We should also support comma separated inputs.

Reviewed By: dkgi

Differential Revision: D28135192

fbshipit-source-id: 7a598b80bac3b08020f5a0e4e5b73a3aba50ffeb