Define rage behavior when no server exists
Summary:
We want to be able to run `pyre rage` after a server crashes, or
after a `pyre stop` command has been run: in this case we should
report that there's no server, and we should still show logs if they
exist.

This commit makes that change.

*Details*

It's actually pretty weird what happens when the server is dead.

I don't think that `rageCommand` depends *directly* on the server at all, it
certainly isn't making any rpc calls. But for some reason the analysis
directory always disappears when the server is dead - so a dead server actually
manifests as file errors on the directory, and that's why this commit mostly
deals with directory handling.

I don't fully grasp why this is the behavior we get, but at any rate this
change seems to make `pyre rage` behave better.

Reviewed By: grievejia

Differential Revision: D28103707

fbshipit-source-id: b6422013f3a993f395ffbda83ece97fa91748d73