Hide class AST implementation detail
Summary:
After doing a code search, I realized that we never call
`explicitly_assigned_attributes` on `Class.t` outside of the
`Class.AttributeComponents.create` function, which is nested
inside it.

I pretty strongly feel like this means `AttributeComponents` is
the only legitimate way to access class attributes, and therefore
the existence of a function to get explicitly-assigned attributes
from a raw class AST body is an implementation detail that should
live inside of `AttributeCompenents` and not be exposed in any
signatures.

My guess is that the existing code is just a leftover from a time
before the `AttributeComponents` module existed.

Reviewed By: grievejia

Differential Revision: D28151847

fbshipit-source-id: ff30a55a1d4c1d16c0f9efee48c31fc341201df2