Move attributes to classSummary.ml [1]
Summary:
The Ast.Statement module had several inner modules that
weren't actually part of the Statement ast, but were instead
defining semantic logic used to find attributes.

These attributes are *not* actually part of the Ast, they are
computed when a ClassSummary.t is constructed out of the raw
Class.t Ast element, and therefore the attribute logic ought
to live there rather than polluting the basic Ast datat ypes.

This commit moves these modules, along with a couple of functions
which acted on them and were (in my view) that were
inappropriately defined inside the Statement.Class and
Statement.Define module. So now all the "basic" (basic meaning not
annotation-aware) attribute-finding logic lives in one place,
the ClassSummary module

An added bonus is that the module structure actually gets
quite a bit simpler - after the refactor it's much clearer
that these "semantic logic" modules do *not* in fact recursively
interact with the raw Ast modules like Class and Define, but
in fact they can be defined entirely in terms of the Ast
and therefore much more loosely coupled than the old code layout
suggested.  Also the fact that we were able to get rid of `open
Analysis` in `statementTest.ml` is a good sign that this change
is good - something's fishy if you have to open Analysis for an
Ast test.

I think the benefits in terms of coherence are easier to see
by skimming the .mli files than the .ml code - there are a bunch
of big functions I moved around so the diff is hard to skim
but none of the logic actually changed.

Reviewed By: grievejia

Differential Revision: D28185779

fbshipit-source-id: 0ca22ec554643d61c0f13a8b4ebcfd038c76f58a