Move attributes to classSummary.ml [0] prep the module
Summary:
After talking with Jia I'm working on a refactor to move
attribute-finding logic out of Ast.Statement and into
Analysis.ClassSummary, which is where the only use case lives.

This diff is just prepping for the change - which will involve
dumping several new modules into classSummary.ml - by moving
that actual ClassSummary top-level logic into a ClassSummary
module which we then include in the top-level.

Why? My impression is that having an eminymous submodule like this
is a pretty standard way (within pyre) of having clear delimiters around
the "main" module of a large multi-module file; we use this pattern in
most of the environment files as well as in type.ml, it makes navigation
by code-folding easier in big files

I isolated this easy commit because I want to make the next commit -
which moves about 1500 lines of code between files - as easy as I can
to look at.

Reviewed By: grievejia

Differential Revision: D28166358

fbshipit-source-id: 154e914cc68db134121dfe70d23056d38cb5f154