Disallow modeling attributes as callables
Summary: We currently allow to model attributes as callables, using the `def foo(): ...` syntax in pysa models. This is very confusing and actually leads to non-determinism in the analysis, since attribute models are treated as callable models, but attributes are not part of the call graph and dependency graph. To prevent this, we disallow modeling attributes as callables. We can still use the `foo: TaintSource[Test]` syntax for attributes.

Reviewed By: dkgi

Differential Revision: D28169809

fbshipit-source-id: f49b82d387e18498f4abf86885c562653fe7361d