Implement a function to resolve a python symbol
Summary:
We want to generate errors if a `.pysa` file tries to model an attribute, a module or a class by using the `def foo.bar()` syntax.
This diff provides an API to resolve a python symbol into a class, module or attribute. This also tells us if the symbol does not exist.

Reviewed By: grievejia

Differential Revision: D28168197

fbshipit-source-id: 9c7f7152f5fa8e13b70e75aed8c427524e46feae