Move attribute accessors to ClassSummary module
Summary:
Class attributes are only ever accessed via a class summary, not
via a ClassAttributes.t "floating around" on it's own.

So to keep the Api easy to use, we ought to move the attributes-fetching
functions up to the ClassSummary level, which means client code doesn't
need to destructure the ClassSummary or access the ClassAttributes module
(which in my opinion should be an implementation detail, not something
client code couples to).

The resulting client callsites are a bit easier to read.

Reviewed By: shannonzhu

Differential Revision: D28255526

fbshipit-source-id: e5cff24ea91d466413ac9ba09855fd5fce2fc757