Push error-handling for infer to outer layer per-define
Summary:
The same untracked-type exceptions that can be thrown by the
`infer_local` code that looks at a function body can also be
thrown by the part of infer that looks at annotations of overridden
parent methods.

So we need to push that exception-handling block to the outer layer
in order to avoid losing mapreduce workers and printing unusable
ocaml-internal stack traces.

Reviewed By: grievejia

Differential Revision: D28259528

fbshipit-source-id: 7d23204d5bde0055c8ccff48ff669a37ec9065a5