Infer import root in pyre init or prompt users
Summary:
We can be smarter in pyre init about where the import root for given source directories are. The presence of an `__init__.py` usually denotes a package root, and indicates imports should be relative to the parent of this file.

Since we only support import roots that are a parent directory of the source root, don't try to do anything fancy if we see the given source directory is ".".

Otherwise, for each source directory given, check if we can find an `__init__.py` file the given directory or any of its parents. Set the import root as the parent of the directory containing the init.

If none is found, let's assume the package root is in a subdirectory of the given source and default to our original behavior. Considered prompting the user to ask for where imports should be relative to here (diff v3), but it wouldn't make sense for the import root to be nested any further down than the given source and we just checked the common case for all parents.

Reviewed By: grievejia

Differential Revision: D28138296

fbshipit-source-id: 43acbdcbff43677be23f55144ff1473fff73061e