Do not read callable models when accessing attributes
Summary:
When accessing an attribute of an instance `Foo.bar`, we are currently reading 2 models: the callable model `Foo.bar` and the attribute model `Foo.bar`. This is leading to non-determinism since this is not actually a call, hence the access to the callable model is not in the call graph / dependency graph. Depending on the number of iterations performed, we might not find a flow.

To fix this, we simply don't read the callable model, as we should. A downside is that we will now miss a few flows when the call graph is imprecise. I will work on this in follow-ups.

Reviewed By: dkgi

Differential Revision: D28268231

fbshipit-source-id: c3f98bd970283bb6068b75c1dfd60dead9119ad8