Propagate errors on automatic server starts
Summary:
When using the pyre API to query information, it will automatically start a pyre server if it is not started already.
The problem is that this implicit start can fail, but the error is silently ignored. It fails later when calling `pyre query`.
To make debugging easier, we should propagate the error on `pyre incremental` right away.

Reviewed By: grievejia

Differential Revision: D28306525

fbshipit-source-id: 56d13afeeb29fdb26cfe8ae9c61c7423d7dfc678