Add an option to limit the maximum trace length
Summary:
Running pysa on large codebases can find very long traces which we are most likely not interested in. This also leads to performance issues.
This diff introduces the `--maximum-trace-length` option, which limits the trace length by a given number.
This can also be used under the `options` set in a `taint.config` file.

Reviewed By: fahndrich

Differential Revision: D28397765

fbshipit-source-id: 356f35f606af77bd424f7112aa6b146cef89b4de