Silence all remaining discarded-qualifier warnings via cast
Summary:
Unfortunately it's not practical to rewrite these two uses of String_val
to respect the `const` qualifier in the return value; we're doing
some explicitly unsafe things here that we know will work by reasoning
but we can't prove it within the constraints of C's typechecker.

The correct way to make the compiler happy here is to cast.

I suppose there's an argument to be made that the warnings are good because
they remind us we're doing unsafe things, but I don't think that's a great
argument for a noisy build... I could add some louder comments though if we
think that's called for.

Reviewed By: grievejia

Differential Revision: D28451150

fbshipit-source-id: f61c6d5a4f1b2fcfabc7169cf312a5a3a31af339