Addition of Fabric RCE/FS Sinks
Summary: Addition of RCE and FS taint sinks for operations found within the various ways to call the Fabric 1.X and 2.X modules. Per discussion with gbleaney, leaving in models that are not currently being identified correctly so that they will be in place once the Pysa team's work on decorators is shipped.

Reviewed By: r0rshark

Differential Revision: D28384888

fbshipit-source-id: de43470525bb74a9750ed89ccb1cb185294aba81