Support using builtins with `builtins.` prefix
Summary:
The rest of the code has special handling of `builtins`, and the module maps to an empty reference. To maintain that view, this fix handles it in preprocessing.

With this fix, we can also handle the case where one is overriding a builtin type, and importing builtins to refer to a builtin type explicitly.

Reviewed By: shannonzhu

Differential Revision: D28428120

fbshipit-source-id: fa58ef818f90f1120b253c68f48cdf1ae0e2d9ef