Demonstrate a bug due to the subtraction optimization when setting a maximum trace length
Summary:
We aggressively prune the taint tree of anything that is present in ancestors.
Because of this, we might end up with a trace length of bottom:
```
(Variable $parameter$x) -> {
  {CallSite(callees=[sink_distance_one], location=maximum_trace_length:64:22-64:23, port=PositionalParameter {x}) -> {Test -> FlowDetails(SimpleFeature: [(SimpleVia "special_sink"), Tito], TraceLength: 2, TitoPosition: [63:15-63:16], LeafName: [LeafName(__test_sink)])}}
  [foo] -> {CallSite(callees=[sink_distance_one], location=maximum_trace_length:64:22-64:23, port=PositionalParameter {x}) -> {Test -> FlowDetails(TitoPosition: [63:18-63:23])}}
   }
```
See how the trace length is not set on the taint for path `[foo]`.
Because of this, the maximum trace length pruning will remove the taint for `x` but not `x.foo`.
Note that in general this is not a problem because the tree shaping will remove `foo` afterward.

Reviewed By: fahndrich

Differential Revision: D28497136

fbshipit-source-id: 19db896a0263f4e62b98f1e985e9b142740ef9a3